+++ platform/RELEASES.md
@@ -87,3 +87,13 @@ Based on:
 - [csharp v2.1.2] platform
 ### Releases
 - [NuGet v2.1.2] https://www.nuget.org/packages/Codat.Platform/2.1.2 - platform
+
+## 2023-09-29 00:19:37
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Platform.yaml
+- Speakeasy CLI 1.91.3 (2.139.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v2.2.0] platform
+### Releases
+- [NuGet v2.2.0] https://www.nuget.org/packages/Codat.Platform/2.2.0 - platform